# openapi.legacy.yaml
openapi: 3.0.3
info:
  title: Rimcast
  version: 0.1.0
  description: Latency-aware edge inference placement for real-time IoT and physical-world AI.
  x-product-slug: ai-decentralized-rimcast
  x-source: ai-in-decentralized+ai/decentralizedai-iotconf-190117170353/
servers:
- url: https://api.example.com
...
paths:
...
  /v1/telemetry/sla-samples:
    post:
      operationId: ingestSlaSample
      tags:
      - Telemetry
      security:
      - ApiKeyAuth: []
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelopeSlaSample'
        '401':
          $ref: '#/components/responses/Unauthorized'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlaSample'
    get:
      operationId: listSlaSamples
      tags:
      - Telemetry
      security:
      - BearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelopeSlaSample'
        '401':
          $ref: '#/components/responses/Unauthorized'
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - name: siteId
        in: query
        schema:
          type: string
...
  /v1/governance/site-freezes:
    post:
      operationId: freezeSitePlacements
      tags:
      - Governance
      security:
      - BearerAuth: []
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelopeSiteFreeze'
        '401':
          $ref: '#/components/responses/Unauthorized'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteFreeze'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    ResponseMeta:
      type: object
      required:
      - requestId
      - timestamp
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    Problem:
      type: object
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
...
    SlaSample:
      type: object
      required:
      - id
      - placementId
      - latencyMs
      properties:
        id:
          type: string
        placementId:
          type: string
        latencyMs:
          type: integer
        withinBudget:
          type: boolean
...
    SiteFreeze:
      type: object
      required:
      - id
      - siteId
      - active
      properties:
        id:
          type: string
        siteId:
          type: string
        active:
          type: boolean
        reason:
          type: string
...
    DataEnvelopeSlaSample:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          $ref: '#/components/schemas/SlaSample'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ListEnvelopeSlaSample:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          type: object
          required:
          - items
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/SlaSample'
            nextCursor:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    DataEnvelopeSiteFreeze:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          $ref: '#/components/schemas/SiteFreeze'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...