# platform/api-server/src/domains/governance/handlers/site-freeze.handlers.ts
/**
 * Handlers for SiteFreeze
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: governance
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { GovernanceDomainModule } from "../dependencies/governance-ddd.dependencies.js";
import type { FreezeSitePlacementsInput, ListSiteFreezesInput } from "@rimcast/services/governance";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/governance/site-freezes
 * Operation: listSiteFreezes
 */
export async function listSiteFreezes(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], siteId?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: GovernanceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListSiteFreezesInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.siteFreezes.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/governance/site-freezes
 * Operation: freezeSitePlacements
 */
export async function freezeSitePlacements(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: GovernanceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: FreezeSitePlacementsInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.siteFreezes.get.execute(input);
  return reply.code(201).send(result);
}