# platform/api-server/src/domains/telemetry/handlers/sla-sample.handlers.ts
/**
 * Handlers for SlaSample
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: telemetry
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { TelemetryDomainModule } from "../dependencies/telemetry-ddd.dependencies.js";
import type { IngestSlaSampleInput, ListSlaSamplesInput } from "@rimcast/services/telemetry";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/telemetry/sla-samples
 * Operation: listSlaSamples
 */
export async function listSlaSamples(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], siteId?: string | string[], modelVersionId?: string | string[], deviceClass?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: TelemetryDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListSlaSamplesInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.slaSamples.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/telemetry/sla-samples
 * Operation: ingestSlaSample
 */
export async function ingestSlaSample(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: TelemetryDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: IngestSlaSampleInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.slaSamples.get.execute(input);
  return reply.code(201).send(result);
}